Moved validator to traits and removed from starting materials and samples
